Call hierarchy doesn't work for Java 11 java.util.Timer.sched #332
During the call hierarchy computation of Timer.sched() from JRE 11,
JavaSearchScope.encloses() is called with the following values:

path=java.base|java/util/Timer.class
enclosingPath=java/util

The check then fails, since the provided enclosing path does not enclose
the path of the type - its missing the module name.

This change adjusts JavaSearchScope to remove the module part of the
path, before checking whether the path is enclosed. In addition, a
warning is logged if JavaSearchScope.encloses(String, String, int) is
reached with a path that contains the symbol '|'.

Fixes: #332