No callers for method found in anonymous types in a jar #375
Changes added for bug
https://bugs.eclipse.org/bugs/show_bug.cgi?id=468127 can use the wrong
anonymous type .class file (in particular, always the "$1" file is
used). This can result in wrong call hierarchy results / wrong method
reference search results, since visited methods are not found in the
wrong .class file and so the search doesn't visit the the bodies of such
methods.

This change moves the fix for bug 468127 to
MatchLocator.reportMatching() for method declarations, adjusting it to
incrementally go up an anonymous types nested hierarchy to find the
method declaration.

A test for the bug is also added.

Fixes: #375

Signed-off-by: Simeon Andreev <simeon.danailov.andreev@gmail.com>